requirements.txt
Django==4.2.1
tox==4.5.1

# Testing
pytest-django==4.5.2
model_bakery==1.11.0
pytest-cov==4.0.0